requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
python-dotenv==1.0.0
pydantic>=2.0.0
email-validator>=2.0.0
httpx>=0.24.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# Real-time and WebRTC signalling
python-socketio>=5.8.0
python-engineio>=4.3.0

# File uploads / static files
python-multipart>=0.0.6
aiofiles>=23.1.0

# Native bcrypt (helpful for passlib on some platforms)
bcrypt>=4.0.1

# Development / debugging helpers (optional)
uvloop>=0.17.0

